Move make_test_inputs to utils file
