+++ src/KroneckerProductKernels.jl
@@ -5,12 +5,13 @@ using LinearAlgebra, Random
 
 using Kronecker: ⊗
 
-import AbstractGPs: MeanFunction, FiniteGP
+import AbstractGPs: AbstractGP, MeanFunction, FiniteGP
 import Distributions, Random
 
-export KernelKroneckerProduct
+export KernelKroneckerProduct, KroneckerFiniteGP
 
 include("kernel.jl")
 include("gp.jl")
+include("utils.jl")
 
 end
